+++ mc.cpp
@@ -48,7 +48,7 @@ void Simulation::mc_sweep() {
 void Simulation::mc_translate() {
     int rand_i = RANDINT(0, WATERS.size() + IONS.size() * ION_PROBABILITY_WEIGHT);
     double old_energy_particle_i = energy_of_particle_with_index(rand_i);
-    if (rand_i < WATERS.size())
+    if (rand_i < (int)WATERS.size())
         WATERS[rand_i]->mc_translate();
     else {
         rand_i = WATERS.size() + ((rand_i - WATERS.size()) % ION_PROBABILITY_WEIGHT);
@@ -74,12 +74,12 @@ bool Simulation::mc_accept(int index, double old_energy_particle_i) {
         TOTAL_ENERGY += total_energy_diff;
     else {
         // undo the move if move not accepted
-        (index < WATERS.size()) ? WATERS[index]->undo_move() : IONS[index - WATERS.size()]->undo_move();
+        (index < (int)WATERS.size()) ? WATERS[index]->undo_move() : IONS[index - WATERS.size()]->undo_move();
 
         // reset partial rho_k's
         dcomplex *column;
         int NUM_TOTAL_PARTICLES = WATERS.size() + IONS.size();
-        for (int k = 0; k < K_VECTORS.size(); k++) {
+        for (unsigned int k = 0; k < K_VECTORS.size(); k++) {
             column = RHO_K_VALUES[k];
             column[NUM_TOTAL_PARTICLES] += column[NUM_TOTAL_PARTICLES + 1] - column[index];
             column[index] = column[NUM_TOTAL_PARTICLES + 1];
